requirements.txt
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
requests==2.31.0
soupsieve==2.4.1
urllib3==2.0.3
fake-useragent==1.5.1
opgg.py
selenium
pandas
webdriver-manager
urllib3
tqdm
lxml